requirements.txt
Django==5.0.3
django-ninja==1.1.0
pydantic==2.6.3
